+++ crates/stackable-operator/CHANGELOG.md
@@ -12,7 +12,7 @@ All notable changes to this project will be documented in this file.
   Currently this is `runAsNonRoot: true`, which might cause product Pods to crash and require changes.
 - BREAKING: `PodSecurityContextBuilder::run_as_non_root` now takes a `bool` instead of assuming consumers always want to set it to `true` ([#1205]).
   This is needed to allow users setting it to `false` in case the new `with_stackable_defaults` function sets it to `true`.
-- BREAKING: `SecurityContextBuilder::run_as_root` is now called on the builder, not as a builder-creation function ([#1205]).
+- BREAKING: `SecurityContextBuilder::run_as_root` has been removed ([#1205]).
 
 [#1205]: https://github.com/stackabletech/operator-rs/pull/1205